Move `SnapError` to SDK
