+++ packages/examples/packages/bip32/snap.manifest.json
@@ -7,7 +7,11 @@
     "url": "https://github.com/MetaMask/snaps.git"
   },
   "source": {
+<<<<<<< HEAD
     "shasum": "7z2/4sDVSx8+xUi9dVoD9v/o10rIhzxLfv7G1pfxGFw=",
+=======
+    "shasum": "Egmb/t82FKMLF/jNE1JomXqC05O1l0Cu5AcFsrEJG40=",
+>>>>>>> 93ebf4b1 (Move `SnapError` to SDK)
     "location": {
       "npm": {
         "filePath": "dist/bundle.js",
