+++ packages/snaps-sdk/package.json
@@ -47,6 +47,7 @@
     "@metamask/eslint-config-jest": "^12.1.0",
     "@metamask/eslint-config-nodejs": "^12.1.0",
     "@metamask/eslint-config-typescript": "^12.1.0",
+    "@metamask/rpc-errors": "^6.1.0",
     "@swc/cli": "^0.1.62",
     "@swc/core": "1.3.78",
     "@swc/jest": "^0.2.26",
