+++ packages/snaps-sdk/src/index.ts
@@ -1,4 +1,11 @@
 // Only internals that are used by other Snaps packages should be exported here.
 export type { EnumToUnion } from './internals';
+export {
+  getErrorData,
+  getErrorMessage,
+  getErrorStack,
+  SNAP_ERROR_CODE,
+  SNAP_ERROR_MESSAGE,
+} from './internals';
 
 export * from './types';
