+++ packages/snaps-sdk/src/internals/index.ts
@@ -1,1 +1,2 @@
+export * from './errors';
 export * from './helpers';
