+++ packages/snaps-types/package.json
@@ -35,7 +35,6 @@
     "lint:dependencies": "depcheck"
   },
   "dependencies": {
-    "@metamask/snaps-utils": "workspace:^",
     "@metamask/utils": "^8.1.0"
   },
   "devDependencies": {
