+++ packages/snaps-types/src/types.ts
@@ -1,3 +1,2 @@
 // Exported again for convenience.
 export type { Json, JsonRpcRequest } from '@metamask/utils';
-export { SnapError } from '@metamask/snaps-utils';
