+++ packages/snaps-utils/coverage.json
@@ -1,6 +1,6 @@
 {
-  "branches": 96.11,
-  "functions": 98.1,
-  "lines": 98.48,
-  "statements": 95.44
+  "branches": 95.83,
+  "functions": 98.99,
+  "lines": 98.65,
+  "statements": 95.49
 }
