+++ packages/snaps-utils/src/errors.test.ts
@@ -1,92 +1,19 @@
 import { errorCodes, JsonRpcError, rpcErrors } from '@metamask/rpc-errors';
+import {
+  SnapError,
+  SNAP_ERROR_CODE,
+  SNAP_ERROR_MESSAGE,
+} from '@metamask/snaps-sdk';
 
 import {
-  getErrorCode,
-  getErrorData,
-  getErrorMessage,
-  getErrorStack,
   isSerializedSnapError,
   isSnapError,
   isWrappedSnapError,
-  SNAP_ERROR_CODE,
-  SNAP_ERROR_MESSAGE,
   SNAP_ERROR_WRAPPER_CODE,
   SNAP_ERROR_WRAPPER_MESSAGE,
-  SnapError,
   unwrapError,
   WrappedSnapError,
 } from './errors';
-
-describe('getErrorMessage', () => {
-  it('returns the error message if the error is an object with a message property', () => {
-    expect(getErrorMessage(new Error('foo'))).toBe('foo');
-    expect(getErrorMessage({ message: 'foo' })).toBe('foo');
-    expect(getErrorMessage(rpcErrors.invalidParams('foo'))).toBe('foo');
-  });
-
-  it('returns the error converted to a string if the error does not have a message property', () => {
-    expect(getErrorMessage('foo')).toBe('foo');
-    expect(getErrorMessage(123)).toBe('123');
-    expect(getErrorMessage(true)).toBe('true');
-    expect(getErrorMessage(null)).toBe('null');
-    expect(getErrorMessage(undefined)).toBe('undefined');
-    expect(getErrorMessage({ foo: 'bar' })).toBe('[object Object]');
-  });
-});
-
-describe('getErrorStack', () => {
-  it('returns the error stack if the error is an object with a stack property', () => {
-    const error = new Error('foo');
-
-    expect(getErrorStack(error)).toBe(error.stack);
-    expect(getErrorStack({ stack: 'foo' })).toBe('foo');
-    expect(getErrorStack(rpcErrors.invalidParams('foo'))).toBeDefined();
-  });
-
-  it('returns undefined if the error does not have a stack property', () => {
-    expect(getErrorStack('foo')).toBeUndefined();
-    expect(getErrorStack(123)).toBeUndefined();
-    expect(getErrorStack(true)).toBeUndefined();
-    expect(getErrorStack(null)).toBeUndefined();
-    expect(getErrorStack(undefined)).toBeUndefined();
-    expect(getErrorStack({ foo: 'bar' })).toBeUndefined();
-  });
-});
-
-describe('getErrorCode', () => {
-  it('returns the error code if the error is an object with a code property', () => {
-    expect(getErrorCode({ code: 123 })).toBe(123);
-    expect(getErrorCode(rpcErrors.invalidParams('foo'))).toBe(-32602);
-  });
-
-  it('returns `errorCodes.rpc.internal` if the error does not have a code property', () => {
-    expect(getErrorCode('foo')).toBe(errorCodes.rpc.internal);
-    expect(getErrorCode(123)).toBe(errorCodes.rpc.internal);
-    expect(getErrorCode(true)).toBe(errorCodes.rpc.internal);
-    expect(getErrorCode(null)).toBe(errorCodes.rpc.internal);
-    expect(getErrorCode(undefined)).toBe(errorCodes.rpc.internal);
-    expect(getErrorCode({ foo: 'bar' })).toBe(errorCodes.rpc.internal);
-  });
-});
-
-describe('getErrorData', () => {
-  it('returns the error data if the error is an object with a data property', () => {
-    expect(getErrorData({ data: { foo: 'bar' } })).toStrictEqual({
-      foo: 'bar',
-    });
-
-    expect(getErrorData(rpcErrors.invalidParams('foo'))).toStrictEqual({});
-  });
-
-  it('returns an empty object if the error does not have a data property', () => {
-    expect(getErrorData('foo')).toStrictEqual({});
-    expect(getErrorData(123)).toStrictEqual({});
-    expect(getErrorData(true)).toStrictEqual({});
-    expect(getErrorData(null)).toStrictEqual({});
-    expect(getErrorData(undefined)).toStrictEqual({});
-    expect(getErrorData({ foo: 'bar' })).toStrictEqual({});
-  });
-});
 
 describe('WrappedSnapError', () => {
   it('wraps an error', () => {
@@ -183,290 +110,6 @@ describe('WrappedSnapError', () => {
           },
         },
       });
-    });
-  });
-});
-
-describe('SnapError', () => {
-  it('creates an error from a message', () => {
-    const error = new SnapError('foo');
-
-    expect(error).toBeInstanceOf(Error);
-    expect(error).toBeInstanceOf(SnapError);
-    expect(error.message).toBe('foo');
-    expect(error.code).toBe(-32603);
-    expect(error.data).toStrictEqual({});
-    expect(error.stack).toBeDefined();
-    expect(error.toJSON()).toStrictEqual({
-      code: -31002,
-      message: 'Snap Error',
-      data: {
-        cause: {
-          code: -32603,
-          message: 'foo',
-          stack: error.stack,
-          data: {},
-        },
-      },
-    });
-  });
-
-  it('creates an error from a message and code', () => {
-    const error = new SnapError({
-      message: 'foo',
-      code: -32000,
-    });
-
-    expect(error).toBeInstanceOf(Error);
-    expect(error).toBeInstanceOf(SnapError);
-    expect(error.message).toBe('foo');
-    expect(error.code).toBe(-32000);
-    expect(error.data).toStrictEqual({});
-    expect(error.stack).toBeDefined();
-    expect(error.toJSON()).toStrictEqual({
-      code: -31002,
-      message: 'Snap Error',
-      data: {
-        cause: {
-          code: -32000,
-          message: 'foo',
-          stack: error.stack,
-          data: {},
-        },
-      },
-    });
-  });
-
-  it('creates an error from a message and data', () => {
-    const error = new SnapError('foo', { foo: 'bar' });
-
-    expect(error).toBeInstanceOf(Error);
-    expect(error).toBeInstanceOf(SnapError);
-    expect(error.message).toBe('foo');
-    expect(error.code).toBe(-32603);
-    expect(error.data).toStrictEqual({ foo: 'bar' });
-    expect(error.stack).toBeDefined();
-  });
-
-  it('creates an error from a message, code, and data', () => {
-    const error = new SnapError(
-      {
-        message: 'foo',
-        code: -32000,
-      },
-      { foo: 'bar' },
-    );
-
-    expect(error).toBeInstanceOf(Error);
-    expect(error).toBeInstanceOf(SnapError);
-    expect(error.message).toBe('foo');
-    expect(error.code).toBe(-32000);
-    expect(error.data).toStrictEqual({ foo: 'bar' });
-    expect(error.stack).toBeDefined();
-    expect(error.toJSON()).toStrictEqual({
-      code: -31002,
-      message: 'Snap Error',
-      data: {
-        cause: {
-          code: -32000,
-          message: 'foo',
-          stack: error.stack,
-          data: {
-            foo: 'bar',
-          },
-        },
-      },
-    });
-  });
-
-  it('creates an error from an error', () => {
-    const error = new SnapError(new Error('foo'));
-
-    expect(error).toBeInstanceOf(Error);
-    expect(error).toBeInstanceOf(SnapError);
-    expect(error.message).toBe('foo');
-    expect(error.code).toBe(-32603);
-    expect(error.data).toStrictEqual({});
-    expect(error.stack).toBeDefined();
-    expect(error.toJSON()).toStrictEqual({
-      code: -31002,
-      message: 'Snap Error',
-      data: {
-        cause: {
-          code: -32603,
-          message: 'foo',
-          stack: error.stack,
-          data: {},
-        },
-      },
-    });
-  });
-
-  it('creates an error from an error and data', () => {
-    const error = new SnapError(new Error('foo'), { foo: 'bar' });
-
-    expect(error).toBeInstanceOf(Error);
-    expect(error).toBeInstanceOf(SnapError);
-    expect(error.message).toBe('foo');
-    expect(error.code).toBe(-32603);
-    expect(error.data).toStrictEqual({ foo: 'bar' });
-    expect(error.stack).toBeDefined();
-    expect(error.toJSON()).toStrictEqual({
-      code: -31002,
-      message: 'Snap Error',
-      data: {
-        cause: {
-          code: -32603,
-          message: 'foo',
-          stack: error.stack,
-          data: {
-            foo: 'bar',
-          },
-        },
-      },
-    });
-  });
-
-  it('creates an error from a JsonRpcError', () => {
-    const error = new SnapError(rpcErrors.invalidParams('foo'));
-
-    expect(error).toBeInstanceOf(Error);
-    expect(error).toBeInstanceOf(SnapError);
-    expect(error.message).toBe('foo');
-    expect(error.code).toBe(-32602);
-    expect(error.data).toStrictEqual({});
-    expect(error.stack).toBeDefined();
-    expect(error.toJSON()).toStrictEqual({
-      code: -31002,
-      message: 'Snap Error',
-      data: {
-        cause: {
-          code: -32602,
-          message: 'foo',
-          stack: error.stack,
-          data: {},
-        },
-      },
-    });
-  });
-
-  it('creates an error from a JsonRpcError and data', () => {
-    const error = new SnapError(rpcErrors.invalidParams('foo'), {
-      foo: 'bar',
-    });
-
-    expect(error).toBeInstanceOf(Error);
-    expect(error).toBeInstanceOf(SnapError);
-    expect(error.message).toBe('foo');
-    expect(error.code).toBe(-32602);
-    expect(error.data).toStrictEqual({ foo: 'bar' });
-    expect(error.stack).toBeDefined();
-    expect(error.toJSON()).toStrictEqual({
-      code: -31002,
-      message: 'Snap Error',
-      data: {
-        cause: {
-          code: -32602,
-          message: 'foo',
-          stack: error.stack,
-          data: {
-            foo: 'bar',
-          },
-        },
-      },
-    });
-  });
-
-  it('creates an error from a JsonRpcError with a code of 0', () => {
-    const error = new SnapError({
-      message: 'foo',
-      code: 0,
-    });
-
-    expect(error).toBeInstanceOf(Error);
-    expect(error).toBeInstanceOf(SnapError);
-    expect(error.message).toBe('foo');
-    expect(error.code).toBe(0);
-    expect(error.data).toStrictEqual({});
-    expect(error.stack).toBeDefined();
-    expect(error.toJSON()).toStrictEqual({
-      code: -31002,
-      message: 'Snap Error',
-      data: {
-        cause: {
-          code: 0,
-          message: 'foo',
-          stack: error.stack,
-          data: {},
-        },
-      },
-    });
-  });
-
-  it('creates an error from a JsonRpcError with a code of 0 and data', () => {
-    const error = new SnapError(
-      {
-        message: 'foo',
-        code: 0,
-      },
-      { foo: 'bar' },
-    );
-
-    expect(error).toBeInstanceOf(Error);
-    expect(error).toBeInstanceOf(SnapError);
-    expect(error.message).toBe('foo');
-    expect(error.code).toBe(0);
-    expect(error.data).toStrictEqual({ foo: 'bar' });
-    expect(error.stack).toBeDefined();
-    expect(error.toJSON()).toStrictEqual({
-      code: -31002,
-      message: 'Snap Error',
-      data: {
-        cause: {
-          code: 0,
-          message: 'foo',
-          stack: error.stack,
-          data: {
-            foo: 'bar',
-          },
-        },
-      },
-    });
-  });
-
-  it('creates an error from a JsonRpcError with a code of 0 and merges the data', () => {
-    const error = new SnapError(
-      {
-        message: 'foo',
-        code: 0,
-        data: {
-          foo: 'baz',
-          bar: 'qux',
-        },
-      },
-      { foo: 'bar' },
-    );
-
-    expect(error).toBeInstanceOf(Error);
-    expect(error).toBeInstanceOf(SnapError);
-    expect(error.message).toBe('foo');
-    expect(error.code).toBe(0);
-    expect(error.data).toStrictEqual({ foo: 'bar', bar: 'qux' });
-    expect(error.stack).toBeDefined();
-    expect(error.toJSON()).toStrictEqual({
-      code: -31002,
-      message: 'Snap Error',
-      data: {
-        cause: {
-          code: 0,
-          message: 'foo',
-          stack: error.stack,
-          data: {
-            foo: 'bar',
-            bar: 'qux',
-          },
-        },
-      },
     });
   });
 });
