+++ packages/snaps-webpack-plugin/package.json
@@ -39,6 +39,7 @@
     "lint:dependencies": "depcheck"
   },
   "dependencies": {
+    "@metamask/snaps-sdk": "workspace:^",
     "@metamask/snaps-utils": "workspace:^",
     "@metamask/utils": "^8.1.0",
     "webpack-sources": "^3.2.3"
