+++ packages/snaps-webpack-plugin/tsconfig.build.json
@@ -14,6 +14,9 @@
   ],
   "references": [
     {
+      "path": "../snaps-sdk/tsconfig.build.json"
+    },
+    {
       "path": "../snaps-utils/tsconfig.build.json"
     }
   ]
